[SID:3156] i18n 키 파서 이원화 해소 — 공유 모듈 하나로 일원화
## 배경
카디르 QA 발견(3558 QA 중) — `scripts/check-i18n-keys.js`(#2371 CI 가드)와
`apps/web/src/lib/i18n-key-coverage.test.ts`(#3111계 vitest)가 같은 판정(코드가
쓰는 i18n 키가 messages에 있나)을 서로 다른 독립 regex 구현으로 수행해 왔다.
#3149(PR#3558)에서 실증: 멤버접근 오귀속 결함(`acc.t()`를 로컬 t로 셈)을 한쪽만
고치자 다른 쪽이 같은 오탐으로 CI를 계속 붉혔다.

## 처방
- `scripts/i18n-key-parser.js`(신규, CJS) — stripComments(#3023 정규식-리터럴
  백틱 방어 포함)·extractHookBindings(useTranslations/getTranslations, const만·
  ㉣)·extractKeyUsages(멤버접근 제외 룩비하인드, #3149 ㉥)·flatten을 한 곳으로.
  plain CJS인 이유: check-i18n-keys.js가 `node`로 직접 실행돼(로더 없음) TS로
  옮기면 CI 스크립트에 컴파일 스텝이 새로 생긴다.
- `check-i18n-keys.js` — 위 4개를 require, 로컬 중복 정의 131줄 제거. DYNAMIC_KEY_
  PREFIXES/isDynamicallyComposed/main()은 이 파일 고유 관심사라 유지. 기존
  module.exports는 그대로(check-i18n-keys.test.js 무변경으로 호환).
- `i18n-key-coverage.test.ts` — 로컬 stripComments/ASSIGN_RE/extractKeyUsages
  74줄 제거, 공유 모듈 import로 교체(`../../../../scripts/i18n-key-parser.js` —
  apps/web tsconfig가 이미 `@sprintable/*`로 워크스페이스 밖 크로스-임포트를
  쓰고 있어 같은 패턴). **부수 효과**: 이전엔 `getTranslations()`(서버 컴포넌트)
  를 못 봤는데 이제 check-i18n-keys.js와 동일하게 잡는다 — 재스캔해도 missing
  0건으로 두 도구가 정확히 합의(일원화의 실증).
- `scripts/i18n-key-parser.test.js`(신규) — 양쪽에 중복돼 있던 extractKeyUsages
  회귀가드(멤버접근 제외 4건)를 정본으로 통합 + stripComments(#3023 백틱 방어·
  #2164 주석-예시 오탐 방지)·extractHookBindings 신규 커버리지 추가.

## AC3(파서 결함 수리는 이제 «한 번») 구조 확認
i18n-key-coverage.test.ts에서 로컬 파서 정의가 물리적으로 사라졌다 — 다음에
파서 버그가 나오면 고칠 파일이 i18n-key-parser.js 하나뿐이라 "한쪽만 고치고
다른 쪽을 잊는" 시나리오 자체가 성립하지 않는다.

## 검증
- `node scripts/check-i18n-keys.js` exit 0(missing 0건, 리팩터 前과 동일).
- root `pnpm vitest run` — **547 files·5097 tests all passed**(apps/web+scripts+
  packages 전체, check-i18n-keys.test.js 무변경으로 green 확認 포함).
- apps/web tsc --noEmit 클린(크로스-워크스페이스 import 문제 없음) + eslint
  타깃 파일 0 warning + `next build` exit 0.
